server/ml_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
together==1.5.8
Pillow==11.2.1
requests==2.32.3
python-dotenv==1.0.0 